+++ backend/server/restservice/leases.go
@@ -79,7 +79,7 @@ func (r *RestAPI) GetLeases(ctx context.Context, params dhcp.GetLeasesParams) mi
 	for i := range keaLeases {
 		l := keaLeases[i]
 		daemonLabel := l.Daemon.GetLabel()
-		cltt := int64(l.CLTT)
+		cltt := l.CLTT
 		state := int64(l.State)
 		localSubnetID := int64(l.LocalSubnetID)
 		validLifetime := int64(l.ValidLifetime)
